+++ src/Input.cpp
@@ -2,7 +2,7 @@
 #include "Device.h"
 using namespace gb::render;
 
-void Input::Process(const std::int32_t code, const bool down)
+void Input::Process(const std::size_t code, const bool down)
 {
 	_CBs.Trigger(code, down);
 	if (code == GB_RENDER_INPUT_MS_L)
